requirements.txt
aiogram==3.15.0
aiohttp==3.10.5
python-dotenv